Revert 2 last commits
